Speed Control tuning (#413)
The PID tuning has been done for R48
The PAS set and tested using Speed Control, but the VC Layer code needs improvment

+++ src/GNR2_Firmware/revA/grn2_ra6t2_ftexboard/FTEX/inc/middleware_motor/pid_regulator.h
@@ -25,8 +25,8 @@ typedef struct PID_Handle
   int32_t   wLowerIntegralLimit;  /**< Lower limit used to saturate the integral
                                        term given by Ki / Ki divisor * integral of
                                        process variable error */
-  int16_t   hUpperOutputLimit;    /**< Upper limit used to saturate the PI output */
-  int16_t   hLowerOutputLimit;    /**< Lower limit used to saturate the PI output */
+  int32_t   hUpperOutputLimit;    /**< Upper limit used to saturate the PI output */
+  int32_t   hLowerOutputLimit;    /**< Lower limit used to saturate the PI output */
   uint16_t  hKpDivisor;           /**< Kp gain divisor, used in conjuction with
                                        Kp gain allows obtaining fractional values.
                                        If FULL_MISRA_C_COMPLIANCY is not defined
